+++ SystemTextJsonPatch/Internal/JSonObjectAdapter.cs
@@ -81,8 +81,7 @@ public sealed class JSonObjectAdapter : IAdapter
 				return false;
 			}
 		}
-		else if (!string.Equals(JsonSerializer.SerializeToNode(currentValue)?.ToString(), JsonSerializer.SerializeToNode(value)?.ToString(),
-			         StringComparison.Ordinal))
+		else if (!ByteHelper.BytesEquals(JsonSerializer.SerializeToUtf8Bytes(currentValue), JsonSerializer.SerializeToUtf8Bytes(value)))
 		{
 			errorMessage = Resources.FormatValueNotEqualToTestValue(JsonSerializer.SerializeToNode(currentValue)?.ToString(), value, segment);
 			return false;