+++ SystemTextJsonPatch/Internal/ListAdapter.cs
@@ -163,8 +163,7 @@ public sealed class ListAdapter : IAdapter
 				return false;
 			}
 		}
-		else if (!string.Equals(JsonSerializer.SerializeToNode(currentValue)?.ToString(), JsonSerializer.SerializeToNode(convertedValue)?.ToString(),
-			         StringComparison.Ordinal))
+		else if (!ByteHelper.BytesEquals(JsonSerializer.SerializeToUtf8Bytes(currentValue), JsonSerializer.SerializeToUtf8Bytes(convertedValue)))
 		{
 			errorMessage = Resources.FormatValueAtListPositionNotEqualToTestValue(JsonSerializer.SerializeToNode(currentValue)?.ToString(), value,
 				positionInfo.Index);